squatchcast/requirements.txt
dash>=0.42,<1.0
pandas>=0.24,<1.0
python-dotenv>=0.10.0,<1.0
palettable>=3.1.1,<4
toolz<1.0